backend/requirements.txt
# FastAPI и зависимости
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
hiredis==2.3.2

# Pydantic для валидации
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Celery для фоновых задач
celery==5.3.6
flower==2.0.1

# S3 Storage
boto3==1.34.28
minio==7.2.3

# Elasticsearch
elasticsearch==8.11.1

# OpenAI и AI
openai==1.10.0
anthropic==0.18.0
pillow==10.2.0
opencv-python==4.9.0.80
ultralytics==8.1.18
torch==2.2.0
torchvision==0.17.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1
geopy==2.4.1

# PDF Generation
reportlab==4.0.9
python-docx==1.1.0

# Image processing
pillow==10.2.0

# OCR для распознавания текста
pytesseract==0.3.10

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Development
python-dotenv==1.0.1

# Monitoring
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0

# Code quality
flake8==7.0.0
black==24.1.1
isort==5.13.2
mypy==1.8.0
pytest-cov==4.1.0
